clippy: drop redundant casts and identity ops in AV1 code
- av1.rs: remove unnecessary 'as u8' on frame_presentation_time /
  buffer_removal_time bit lengths (expressions are already u8)
- OBU extension-byte idc parsing (av1_inline test, dump_stream,
  vulkan access_unit, vaapi decoder, nvdec av1): '(e >> 0) & 0x1f'
  -> 'e & 0x1f'

Behavior unchanged: workspace tests 358/358, AV1 matrix 12/12 @ 300/300.

+++ crates/vacc-parser/src/av1.rs
@@ -1150,7 +1150,7 @@ impl Av1Parser {
             // the decoder model is present and pictures are not equally spaced.
             if sps.decoder_model_info_present_flag && !sps.equal_picture_interval {
                 let _frame_presentation_time =
-                    r.read_bits((sps.frame_presentation_time_length_minus_1 + 1) as u8)?;
+                    r.read_bits(sps.frame_presentation_time_length_minus_1 + 1)?;
             }
             // Spec: display_frame_id f(idLen) when frame ID numbers are used.
             if sps.frame_id_numbers_present_flag {
@@ -1178,7 +1178,7 @@ impl Av1Parser {
             && !sps.equal_picture_interval
         {
             let _frame_presentation_time =
-                r.read_bits((sps.frame_presentation_time_length_minus_1 + 1) as u8)?;
+                r.read_bits(sps.frame_presentation_time_length_minus_1 + 1)?;
         }
 
         fh.showable_frame = if show_frame {
@@ -1314,7 +1314,7 @@ impl Av1Parser {
         if sps.decoder_model_info_present_flag {
             let brt_present = r.read_bit()?;
             if brt_present {
-                let n = (sps.buffer_removal_time_length_minus_1 + 1) as u8;
+                let n = sps.buffer_removal_time_length_minus_1 + 1;
                 for op in 0..=sps.operating_points_cnt_minus_1 {
                     let op = op as usize;
                     if !sps.decoder_model_present_for_this_op[op] {
